deleting images from server while removing the product from database

+++ MERN/ecommerce-sample/controllers/admin.controller.js
@@ -1,5 +1,6 @@
 const Product = require('./../models/product.model');
 const { validationResult } = require('express-validator');
+const { deleteFile } = require('./../util/file');
 
 exports.getAddProduct = (req, res, next) => {
     res.render('admin/edit-product', {
@@ -19,11 +20,23 @@ exports.postAddProduct = (req, res, next) => {
     let errorMessage = '';
     const errors = validationResult(req);
     errors.array().forEach(err => {
-        errorMessage = errorMessage + err.msg;
+        errorMessage = errorMessage + ' ' + err.msg;
     });
+    const image = req.file;
+    console.log("image", image);
 
     // validation 
     if (!errors.isEmpty()) {
+        if (image) {
+            // since error in validation occurs
+            // and if we have uploaded the image
+            // except for the case of unsupported file types
+            // the image gets saved in the server even though we may have errors
+            // in other input fields
+            // so we delete such files
+            // console.log("error in validation but image is saved");
+            deleteFile(image.path);
+        }
         return res.status(422).render('admin/edit-product', {
             pageTitle: 'Add Product',
             path: 'add-product',
@@ -38,9 +51,6 @@ exports.postAddProduct = (req, res, next) => {
             }
         });
     }
-
-    const image = req.file;
-    console.log("image", image);
     // checking if image is valid/supported file types
     if (!image) {
         return res.status(422).render('admin/edit-product', {
@@ -129,10 +139,24 @@ exports.postUpdateProduct = (req, res, next) => {
     let errorMessage = '';
     const errors = validationResult(req);
     errors.array().forEach(err => {
-        errorMessage = errorMessage + err.msg;
+        errorMessage = errorMessage + ' ' + err.msg;
     });
+
+    // checking if the image is sent in the request
+    // undefined if no image is uploaded
+    const image = req.file;
     // validation 
     if (!errors.isEmpty()) {
+        if (image) {
+            // since error in validation occurs
+            // and if we have uploaded the image
+            // except for the case of unsupported file types
+            // the image gets saved in the server even though we may have errors
+            // in other input fields
+            // so we delete such files
+            // console.log("error in validation but image is saved");
+            deleteFile(image.path);
+        }
         return res.status(422).render('admin/edit-product', {
             pageTitle: 'Edit Product',
             path: 'edit-product',
@@ -149,10 +173,6 @@ exports.postUpdateProduct = (req, res, next) => {
         });
     }
 
-    // checking if the image is sent in the request
-    // undefined if no image is uploaded
-    const image = req.file;
-
     Product.findById(req.body.id)
         .then(product => {
             // console.log("product userId and req.user._id", product.userId, req.user._id);
@@ -163,6 +183,11 @@ exports.postUpdateProduct = (req, res, next) => {
             product.price = req.body.price;
             product.description = req.body.description;
             if (image) {
+                // we got an image in the request body so
+                //  we are sure that we want to replace the file
+                // so we first delete the file
+                deleteFile(product.imageUrl.slice(1));
+                // set the image path in database
                 product.imageUrl = '/' + image.path;
             }
             // console.log("product to update>>>", product);
@@ -183,7 +208,13 @@ exports.postUpdateProduct = (req, res, next) => {
 exports.postDeleteProduct = (req, res, next) => {
 
     const id = req.body.id;
-    Product.deleteOne({ _id: id, userId: req.user._id })
+    Product.findById(id).then(Product => {
+            if (!product) {
+                return next(new Error('Product not found'));
+            }
+            deleteFile(Product.imageUrl.slice(1));
+            return Product.deleteOne({ _id: id, userId: req.user._id })
+        })
         .then(response => {
             res.redirect("/admin/products");
         })
